// src/app/models/poker-hand/hand-action-type.ts
export enum HandActionType {
    DealCards = 'DealCards',
    Ante = 'Ante',
    SmallBlind = 'SmallBlind',
    BigBlind = 'BigBlind',
    Straddle = 'Straddle',
    Fold = 'Fold',
    Check = 'Check',
    Bet = 'Bet',
    Call = 'Call',
    Raise = 'Raise',
    UncalledBet = 'UncalledBet',
    Flop = 'Flop',
    Turn = 'Turn',
    River = 'River',
    ShowCards = 'ShowCards',
    Showdown = 'Showdown',
}